LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylenonce-blocked-expected.txt

CONSOLE MESSAGE: line 1: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'self'".

CONSOLE MESSAGE: line 6: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'self'".

Style that does not match a 'nonce-*' expression in 'style-src' should not be applied to the page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(el).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
This text should be green.
